# src/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Flask Api Course ",
    "contact": {
      "email": ""
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "Film Request",
      "description": "Example API for requesting and return films requests"
    }
  ],
  "paths": {
    "/films": {
      "get": {
        "tags": [
          "Film Request"
        ],
        "summary": "Returns all available films in the system",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/filmsResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Film Request"
        ],
        "summary": "Create a new film",
        "requestBody": {
          "description": "Film Request Post Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/films"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          }
        }
      }
    },
    "/films/{uuid}": {
      "get": {
        "tags": [
          "Film Request"
        ],
        "summary": "Get film with given UUID",
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Film Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/film"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          },
          "404": {
            "description": "Failed. Film was not found."
          }
        }
      },
      "put": {
        "summary": "edit a film request by uuid",
        "tags": [
          "Film Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Film Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "requestBody": {
          "description": "Film Request Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/film"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/film"
            }
          },
          "400": {
            "description": "Failed. Wrong data."
          }
        }
      },
      "delete": {
        "summary": "Delete Film Request by UUID",
        "tags": [
          "Film Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "uuid",
            "required": true,
            "description": "Film Request uuid",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/uuid"
            }
          },
          "400": {
            "description": "Failed. Misunderstood Request."
          },
          "404": {
            "description": "Failed. Film was not found."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "uuid": {
        "properties": {
          "uuid": {
            "type": "string"
          }
        }
      },
      "films": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "format": "string"
          },
          "release_date": {
            "type": "string",
            "format": "string"
          },
          "distributed_by": {
            "type": "string",
            "format": "string"
          },
          "description": {
            "type": "string",
            "format": "string"
          },
          "rating": {
            "type": "number"
          },
          "length": {
            "type": "number"
          }
        }
      },
      "film": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "format": "string"
          },
          "release_date": {
            "type": "string",
            "format": "string"
          },
          "distributed_by": {
            "type": "string",
            "format": "string"
          },
          "description": {
            "type": "string",
            "format": "string"
          },
          "rating": {
            "type": "number"
          },
          "length": {
            "type": "number"
          }
        }
      },
      "filmsResponse": {
        "type": "object",
        "properties": {
          "bookRequest": {
            "type": "object",
            "additionalProperties": {
              "$ref": "#/components/schemas/film"
            }
          }
        }
      }
    }
  }
}